requirements.txt
flask==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.2
idna==3.10
ImageMagic==0.2.1
itsdangerous==2.2.0
Jinja2==3.1.5
loguru==0.7.3
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
pillow==9.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
schedule==1.2.2
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Wand==0.6.7
watchdog==3.0.0
Werkzeug==3.1.3